+++ pkgs/firehose/lib/src/utils.dart
@@ -108,3 +108,14 @@ bool expectEnv(String? value, String name) {
     return true;
   }
 }
+
+enum Severity {
+  success(':heavy_check_mark:'),
+  info(':heavy_check_mark:'),
+  warning(':warning:'),
+  error(':exclamation:');
+
+  final String emoji;
+
+  const Severity(this.emoji);
+}
